\name{anRichment-package}
\alias{anRichment-package}
\alias{anRichment}
\docType{package}
\title{
Annotation And Enrichemnt Functions
}
\description{
 This package contains functions for creating gene set collections, calculating enrichments of
new gene lists in existing gene set collections, and also provides gene set collections collected from
various online references.
}
\details{
\tabular{ll}{
Package: \tab anRichment\cr
Type: \tab Package\cr
Version: \tab 0.80-1 \cr
Date: \tab 2016-02-19\cr
License: \tab GPL-3
}
%::FIXME:: An overview of how to use the package, including the most important ::FIXME::
%::FIXME:: functions ::FIXME::
}
\author{
Author: Jeremy A. Miller, Peter Langfelder, Jeff Aaronson

Maintainer: Peter Langfelder <Peter.Langfelder@gmail.com>
}
\references{
Some of the brain cell type gene sets from single cell data are described in 

Wang N, Langfelder P, Stricos M, Ramanathan L, Richman JB, Vaca R, Plascencia M, Gu X, Zhang S, Tamai TK, Zhang L, Gao F,
Ouk K, Lu X, Ivanov LV, Vogt TF, Lu QR, Morton AJ, Colwell CS, Aaronson JS, Rosinski J, Horvath S, Yang XW. 

Mapping brain gene coexpression in daytime transcriptomes unveils diurnal molecular networks and deciphers perturbation gene signatures. Neuron. 2022 Oct 19;110(20):3318-3338.e9. doi: 10.1016/j.neuron.2022.09.028. PMID: 36265442; PMCID: PMC9665885.
}
\keyword{ package }
