\name{nDataSets}
\alias{nDataSets}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Number of gene sets in a collection. 
}
\description{
  Returns the number of gene sets in a collection.
}
\usage{
nDataSets(collection)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{collection}{
     Collection object in which to determine the number of gene sets.
}
}
\value{
  Integer giving the number of gene sets.
}
\author{
 Peter Langfelder
}
\keyword{ misc }
