\name{renameGroup}
\alias{renameGroup}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Rename a group within a collection
}
\description{
Rename a group within a collection.
}
\usage{
renameGroup(collection, oldGroup, newGroup)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{collection}{
A collection of gene sets and properties.
}
  \item{oldGroup}{
Name of the group to be renamed.
}
  \item{newGroup}{
New name for the renamed group.
}
}
\value{
The collection with the group renamed.
}
\author{
Peter Langfelder
}
\keyword{misc}% __ONLY ONE__ keyword per line
