\name{anRichment-package}
\alias{anRichment-package}
\alias{anRichment}
\docType{package}
\title{
Collections and other data for annotation and enrichment calculations (anRichmentMethods)
}
\description{
This package contains gene set collections and other data for use with anRichmentMethods.
}
\details{
\tabular{ll}{
Package: \tab anRichment\cr
Type: \tab Package\cr
Version: \tab 0.80-1 \cr
Date: \tab 2016-02-19\cr
License: \tab GPL-3
}
%::FIXME:: An overview of how to use the package, including the most important ::FIXME::
%::FIXME:: functions ::FIXME::
}
\author{
Author: Jeremy A. Miller, Peter Langfelder, Jeff Aaronson

Maintainer: Peter Langfelder <Peter.Langfelder@gmail.com>
}
\references{
::FIXME:: Literature or other references for background information ::FIXME::
}
\keyword{ package }
